Major Update
The FootBallField class was updated to include code for display lists.
Also, the objects in the world were divided into 2 categories:
1) Stationery objects with respect to the camera - Sun, Sky, Ground
2) Moving Objects with respect to the camera - Goal Post

The camera now has the ability to rotate about the Y-axis, by pressing
the 'j' or ' l' keys.

This application can now be opened in full screen, depending on the
value of the macro FULL_SCREEN